add getData on FireBase, add progressBar and sort DESC
