+++ .idea/misc.xml
@@ -3,6 +3,7 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="..\:/AndroidGeek/UI/NewsApp40/app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.1515" />
         <entry key="..\:/AndroidGeek/UI/NewsApp40/app/src/main/res/drawable/background_error.xml" value="0.1" />
         <entry key="..\:/AndroidGeek/UI/NewsApp40/app/src/main/res/drawable/background_normal.xml" value="0.1" />
         <entry key="..\:/AndroidGeek/UI/NewsApp40/app/src/main/res/drawable/ic_arrow_right.xml" value="0.3594871794871795" />
