+++ app/src/main/java/com/geektech/newsapp40/data/room/model/NewsModel.java
@@ -22,6 +22,9 @@ public class NewsModel implements Serializable {
         this.createdAt = createdAt;
         this.description = description;
     }
+    public NewsModel(){
+
+    }
 
     public int getId() {
         return id;
