+++ app/src/main/res/layout/fragment_news.xml
@@ -15,8 +15,8 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_gravity="center"
-        android:layout_marginTop="40dp"
         android:layout_marginHorizontal="24dp"
+        android:layout_marginTop="40dp"
         android:layout_marginBottom="80dp"
         app:boxStrokeColor="@color/black"
         app:errorEnabled="true"
@@ -37,5 +37,13 @@
         android:layout_gravity="center"
         android:paddingVertical="12dp"
         android:text="Save" />
+
+    <ProgressBar
+        android:layout_marginTop="60dp"
+        android:layout_gravity="center"
+        android:id="@+id/progressBar"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:visibility="invisible" />
 
 </LinearLayout>
